% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-bindFuzzyData.R
\name{bindColums}
\alias{bindColums}
\alias{bindColums,FuzzyData,FuzzyData-method}
\title{Function that binds two FuzzyData together into one FuzzyData}
\usage{
bindColums(data1, data2)

\S4method{bindColums}{FuzzyData,FuzzyData}(data1, data2)
}
\arguments{
\item{data1}{An object of \code{\linkS4class{FuzzyData}}.}

\item{data2}{An object of \code{\linkS4class{FuzzyData}}.}
}
\value{
An object of class \code{\linkS4class{FuzzyData}}
}
\description{
This methods construct object \code{\linkS4class{FuzzyData}} based on two \code{\linkS4class{FuzzyData}}.
The functions merges the sources into single output. This method should be used in situations when both
weights and input data are fuzzy.
}
