% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-buildFuzzyPairwiseComparisonMatrix.R
\docType{methods}
\name{buildFuzzyPairwiseComparisonMatrix}
\alias{buildFuzzyPairwiseComparisonMatrix}
\alias{buildFuzzyPairwiseComparisonMatrix,list-method}
\title{Function that builds Fuzzy Pairwise Comparison Matrix based on list of Pairwise Comparison Matrices}
\usage{
buildFuzzyPairwiseComparisonMatrix(listOfMatrices)

\S4method{buildFuzzyPairwiseComparisonMatrix}{list}(listOfMatrices)
}
\arguments{
\item{listOfMatrices}{An object of \code{\linkS4class{list}}.}
}
\value{
An object of class \code{\linkS4class{FuzzyPairwiseComparisonMatrix}}
}
\description{
This functions builds Fuzzy Pairwise Comparison Matrix based on list of Pairwise Comparison Matrices
the resulting Fuzzy Pairwise Comparison Matrix is calculated as minimum, geometric mean and maximum of
each cell of all Pairwise Comparison Matrices in \code{listOfMatrices}.
}
