% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-buildPairwiseComparisonMatrix.R
\name{buildPairwiseComparisonMatrix}
\alias{buildPairwiseComparisonMatrix}
\alias{buildPairwiseComparisonMatrix,list-method}
\title{Function that builds Pairwise Comparison Matrix based on list of Pairwise Comparison Matrices}
\usage{
buildPairwiseComparisonMatrix(listOfMatrices)

\S4method{buildPairwiseComparisonMatrix}{list}(listOfMatrices)
}
\arguments{
\item{listOfMatrices}{An object of \code{\linkS4class{list}}.}
}
\value{
An object of class \code{\linkS4class{PairwiseComparisonMatrix}}
}
\description{
This functions builds Pairwise Comparison Matrix based on list of Pairwise Comparison Matrices
the resulting Pairwise Comparison Matrix is calculated as geometric mean of all
Pairwise Comparison Matrices in \code{listOfMatrices}.
}
