% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-calculateWeights.R
\name{calculate_weighting_vector}
\alias{calculate_weighting_vector}
\alias{calculate_weighting_vector,FuzzyWeights-method}
\alias{calculate_weighting_vector,}
\alias{FuzzyWeights-method}
\title{Function to calculate fuzzy weighting vector}
\usage{
calculate_weighting_vector(fuzzyWeights)

\S4method{calculate_weighting_vector}{FuzzyWeights}(fuzzyWeights)
}
\arguments{
\item{fuzzyWeights}{object of \linkS4class{FuzzyWeights}}
}
\value{
weighting vector for defined \linkS4class{FuzzyWeights}.
}
\description{
This functions calculates fuzzy weighting vector from \code{\linkS4class{FuzzyWeights}}.
The calculation was first described by Chang (1996).
}
\references{
CHANG, Da-Yong, 1996, Applications of the extent analysis method on fuzzy AHP. European Journal of Operational Research. 1996. Vol. 95, no. 3, p. 649–655. DOI 10.1016/0377-2217(95)00300-2.
}
