% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-calculateWeights.R
\docType{methods}
\name{calculateWeights}
\alias{calculateWeights}
\alias{calculateWeights,FuzzyPairwiseComparisonMatrix-method}
\alias{calculateWeights,PairwiseComparisonMatrix-method}
\title{Function to calculate fuzzy weights based on comparison matrix}
\usage{
calculateWeights(comparisonMatrix)

\S4method{calculateWeights}{PairwiseComparisonMatrix}(comparisonMatrix)

\S4method{calculateWeights}{FuzzyPairwiseComparisonMatrix}(comparisonMatrix)
}
\arguments{
\item{comparisonMatrix}{object of either \linkS4class{PairwiseComparisonMatrix} or \linkS4class{FuzzyPairwiseComparisonMatrix}}
}
\description{
This functions calculates \code{\linkS4class{Weights}} or \code{\linkS4class{FuzzyWeights}}
based on input pairwise comparison matrix.
}
\seealso{
\link{PairwiseComparisonMatrix-class}
}

