% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-consistencyIndex.R
\name{consistencyIndex}
\alias{consistencyIndex}
\alias{consistencyIndex,PairwiseComparisonMatrix-method}
\alias{consistencyIndex,FuzzyPairwiseComparisonMatrix-method}
\title{Function to determine Consistency Index}
\usage{
consistencyIndex(comparisonMatrix)

\S4method{consistencyIndex}{PairwiseComparisonMatrix}(comparisonMatrix)

\S4method{consistencyIndex}{FuzzyPairwiseComparisonMatrix}(comparisonMatrix)
}
\arguments{
\item{comparisonMatrix}{A \code{\linkS4class{PairwiseComparisonMatrix}}}
}
\value{
A numeric value of Consistency index.
}
\description{
This methods calculates Consistency index for \code{\linkS4class{PairwiseComparisonMatrix}}.
}
