% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-defuzzify.R
\docType{methods}
\name{defuzziffy}
\alias{defuzziffy}
\alias{defuzziffy,FuzzyData,character-method}
\title{Function to defuzzify fuzzy data}
\usage{
defuzziffy(fuzzyData, type)

\S4method{defuzziffy}{FuzzyData,character}(fuzzyData, type)
}
\arguments{
\item{fuzzyData}{A \code{\linkS4class{FuzzyData}}}

\item{type}{A \code{"character"} representing type of defuzzification. Currently implemented methods are
\code{"Yager"} and default.}
}
\value{
A numeric value of defuzzified value, based on deffuzification method.
}
\description{
This function defuzzifies \code{\linkS4class{FuzzyData}} into single value.
The \code{\linkS4class{FuzzyData}} can only contain one fuzzy number, otherwise the defuzzification is not possible.
}

