% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-FuzzyData.R
\docType{class}
\name{FuzzyData-class}
\alias{FuzzyData-class}
\title{Class "FuzzyData"}
\description{
An S4 class to represent fuzzy data.
}
\section{Slots}{

\describe{
\item{\code{fnMin}}{A numeric vector of minimal values of fuzzy data.}

\item{\code{fnModal}}{A numeric vector of modal values of fuzzy data.}

\item{\code{fnMax}}{A numeric vector of maximal values of fuzzy data.}
}}

