% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-FuzzyData.R
\docType{methods}
\name{fuzzyData}
\alias{fuzzyData}
\alias{fuzzyData,matrix-method}
\title{Function that creates FuzzyData}
\usage{
fuzzyData(data)

\S4method{fuzzyData}{matrix}(data)
}
\arguments{
\item{data}{A \code{matrix} with 3 colums.}
}
\value{
An object of class \code{\linkS4class{FuzzyData}}
}
\description{
This methods construct object \code{\linkS4class{FuzzyData}} based on provided \code{matrix}.
The matrix needs to be have rows represent individual fuzzy numbers and three colums that
represent minimal, modal and maximal value of fuzzy number.
}
\seealso{
\linkS4class{FuzzyData}
}

