% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-FuzzyPairwiseComparisonMatrix.R
\docType{class}
\name{FuzzyPairwiseComparisonMatrix-class}
\alias{FuzzyPairwiseComparisonMatrix-class}
\title{Class "FuzzyPairwiseComparisonMatrix"}
\description{
An S4 class to represent a fuzzy pairwise comparison matrix.
}
\section{Slots}{

\describe{
\item{\code{fnMin}}{A matrix of minimal values of fuzzy preferences.}

\item{\code{fnModal}}{A matrix of modal values of fuzzy preferences.}

\item{\code{fnMax}}{A matrix of maximal values of fuzzy preferences.}

\item{\code{variableNames}}{Names of variables in the pariwise comparison matrix obtained either as colnames or rownames.}
}}

