% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-FuzzyWeights.R
\docType{class}
\name{FuzzyWeights-class}
\alias{FuzzyWeights-class}
\title{Class "FuzzyWeights"}
\description{
An S4 class to represent fuzzy weights for fuzzy AHP calculation.
}
\section{Slots}{

\describe{
\item{\code{fnMin}}{Object of class \code{numeric} containing minimal values of fuzzy weights.}

\item{\code{fnModal}}{Object of class \code{numeric} containing modal values of fuzzy weights.}

\item{\code{fnMax}}{Object of class \code{numeric} containing maximal values of fuzzy weights.}
}}

