% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-getFuzzyNumbers.R
\docType{methods}
\name{getFuzzyNumber}
\alias{getFuzzyNumber}
\alias{getFuzzyNumber,FuzzyData,integer-method}
\alias{getFuzzyNumber,FuzzyWeights,integer-method}
\title{Function to extract specific fuzzy numbers}
\usage{
getFuzzyNumber(object, index)

\S4method{getFuzzyNumber}{FuzzyData,integer}(object, index)

\S4method{getFuzzyNumber}{FuzzyWeights,integer}(object, index)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{FuzzyData}} or \code{\linkS4class{FuzzyWeights}}}

\item{index}{An object of class \code{integer} that represents one or more indices to extract the data from}
}
\value{
A \code{matrix} where rows are fuzzy numbers and columns are important values.
}
\description{
This methods helps with extracting fuzzy numbers from \code{\linkS4class{FuzzyData}} and
\code{\linkS4class{FuzzyWeights}}.
}

