% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-PairwiseComparisonMatrix.R
\docType{class}
\name{PairwiseComparisonMatrix-class}
\alias{PairwiseComparisonMatrix-class}
\title{Class "PairwiseComparisonMatrix"}
\description{
An S4 class to represent a pairwise comparison matrix.
}
\section{Slots}{

\describe{
\item{\code{valuesChar}}{A pairwise comparison matrix based on Saaty's method as characters.}

\item{\code{values}}{A pairwise comparison matrix based on Saaty's method as numeric.}
}}

