% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-textRepresentation.R
\docType{methods}
\name{textRepresentation}
\alias{textRepresentation}
\alias{textRepresentation,FuzzyPairwiseComparisonMatrix-method}
\alias{textRepresentation,FuzzyPairwiseComparisonMatrix,logical-method}
\alias{textRepresentation,PairwiseComparisonMatrix-method}
\alias{PairwiseComparisonMatrix,logical-method}
\title{Function to create text representation of (Fuzzy) Pairwise comparison matrix}
\usage{
textRepresentation(x, whole = TRUE)

\S4method{textRepresentation}{FuzzyPairwiseComparisonMatrix}(x, whole = TRUE)

\S4method{textRepresentation}{PairwiseComparisonMatrix}(x, whole = TRUE)
}
\arguments{
\item{x}{An object of class \code{\linkS4class{FuzzyPairwiseComparisonMatrix}}
or \code{\linkS4class{PairwiseComparisonMatrix}}}

\item{whole}{A boolean object specifying if the whole matrix should be created
or only its upper half should be filled.}
}
\value{
A \code{dataframe} of \code{character} type.
}
\description{
This methods creates text representation of Pairwise comparison matrices.
}
