% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-weakConsistency.R
\docType{methods}
\name{weakConsistency}
\alias{weakConsistency}
\alias{weakConsistency,FuzzyPairwiseComparisonMatrix-method}
\alias{weakConsistency,PairwiseComparisonMatrix-method}
\title{Function to assess weak consistency of Comparison Matrix}
\usage{
weakConsistency(PairwiseComparisonMatrix)


  \S4method{weakConsistency}{FuzzyPairwiseComparisonMatrix}(PairwiseComparisonMatrix)

\S4method{weakConsistency}{PairwiseComparisonMatrix}(PairwiseComparisonMatrix)
}
\arguments{
\item{PairwiseComparisonMatrix}{A \linkS4class{FuzzyPairwiseComparisonMatrix} or
\linkS4class{PairwiseComparisonMatrix}.}
}
\value{
Boolean value indicating if Comparison Matrix passed the weak consistency test and a warning message
listing the problematic triplets if the matrix is not consisten.
}
\description{
Check if for \eqn{a_{ij}>1,a_{jk}>1}{a[i,j]>1,a[j,k]>1} applies that
\eqn{a_{ik}>\max(a_{ij},a_{jk})}{a[i,k]>max(a[i,j],a[j,k])} for all \eqn{i,j,k = 1,2,\dots,n}, where
\eqn{n} is size of \eqn{a}.
}

