% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Weights.R
\docType{class}
\name{Weights-class}
\alias{Weights-class}
\title{Class "Weights"}
\description{
An S4 class representing weights for AHP calculation. Each value in \code{numeric} vector
represents one weight.
}
\section{Slots}{

\describe{
\item{\code{weights}}{Object of class \code{numeric} containing weights.}
}}

