% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-consistencyRatio.R
\docType{methods}
\name{consistencyRatio}
\alias{consistencyRatio}
\alias{consistencyRatio,PairwiseComparisonMatrix-method}
\title{Function to determine Consistency Ratio}
\usage{
consistencyRatio(comparisonMatrix)

\S4method{consistencyRatio}{PairwiseComparisonMatrix}(comparisonMatrix)
}
\arguments{
\item{comparisonMatrix}{A \code{\linkS4class{PairwiseComparisonMatrix}}}
}
\value{
A numeric value of Consistency Ratio, for \code{\linkS4class{PairwiseComparisonMatrix}} with more than 10
rows a value of Consistency index, that needs to be further evaluated.
}
\description{
This methods calculates Consistency Ratio for \code{\linkS4class{PairwiseComparisonMatrix}}.
The consistency ratio can only be provided for \code{\linkS4class{PairwiseComparisonMatrix}} with
less than 10 rows. For bigger matrices the value is not known.
}

