% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-FuzzyData.R
\docType{methods}
\name{fuzzyData}
\alias{fuzzyData}
\alias{fuzzyData,matrix-method}
\title{Function that creates FuzzyData}
\usage{
fuzzyData(data, single.value = TRUE)

\S4method{fuzzyData}{matrix}(data, single.value = TRUE)
}
\arguments{
\item{data}{A \code{matrix} with 3 colums.}

\item{single.value}{An optional boolean parameter (default value TRUE) specifying if the data to be
turn into fuzzy data is single vector of fuzzy numbers (then it needs to have 3 colums) or if the
whole matrix needs to be turn into fuzzy values.}
}
\value{
An object of class \code{\linkS4class{FuzzyData}}
}
\description{
This methods construct object \code{\linkS4class{FuzzyData}} based on provided \code{matrix}.
The matrix needs to be have rows represent individual fuzzy numbers and three colums that
represent minimal, modal and maximal value of fuzzy number.
}
\seealso{
\linkS4class{FuzzyData}
}
